SINGA-246 Imgtool for image augmentation   - put Imgtool class into a separate file   - some improvements
